requirements.txt
Django==5.1.5
django-filter==2.4.0
jsonschema==3.2.0
django-health-check==3.6.1
djangorestframework-simplejwt[crypto]==5.4.0
djangorestframework==3.15.2
coreapi==2.3.3
psycopg==3.2.4
django-oauth-toolkit==1.3.2
futures==3.1.1
pyswagger==0.8.39
bravado-core==5.13.1
drf-yasg==1.21.7
requests==2.32.0
aiohttp==3.10.11
django-cors-headers==3.10.1
gunicorn==22.0.0
voluptuous==0.12.1
